seperate azurefunction from azure as vendor metadata if different
