+++ lib/config/default.js
@@ -493,6 +493,14 @@ defaultConfig.definition = () => ({
       default: true
     },
     /**
+     * This flag dictates whether the agent attempts to read environment variables
+     * and invocation context to get info about the Azure Function called.
+     */
+    detect_azurefunction: {
+      formatter: boolean,
+      default: true
+    },
+    /**
      * This flag dictates whether the agent attempts to read files
      * to get info about the container the process is running in.
      *
