+++ lib/utilization/azure-info.js
@@ -24,21 +24,6 @@ function fetchAzureInfo(agent, callback) {
     return setImmediate(callback, null, results)
   }
 
-  // Detect if Azure Function.
-  // TODO: How to access Azure Function invocation context?
-  if (process.env.REGION_NAME && process.env.WEBSITE_OWNER_NAME && process.env.WEBSITE_SITE_NAME) {
-    // Derive vendors metadata from environment variables.
-    const subscriptionId = process.env.WEBSITE_OWNER_NAME.split('+')[0]
-    const resourceGroupName = process.env.WEBSITE_RESOURCE_GROUP ?? process.env.WEBSITE_OWNER_NAME.match(/\+([a-zA-Z0-9]+)-[a-zA-Z0-9]+(?:-Linux)?/)?.[0]
-    const azureFunctionAppName = process.env.WEBSITE_SITE_NAME
-    results = {
-      'faas.app_name': `/subscriptions/${subscriptionId}/resourceGroups/${resourceGroupName}/providers/Microsoft.Web/sites/${azureFunctionAppName}`,
-      'cloud.region': process.env.REGION_NAME
-    }
-    // Call back!
-    callback(null, results)
-  }
-
   // eslint-disable-next-line sonarjs/no-hardcoded-ip
   const instanceHost = '169.254.169.254'
   const apiVersion = '2017-03-01'
