+++ lib/utilization/index.js
@@ -10,7 +10,7 @@ const logger = require('../logger').child({ component: 'utilization' })
 const VENDOR_METHODS = {
   aws: require('./aws-info'),
   azure: require('./azure-info'),
-  azurefunction: require('./azure-info'),
+  azurefunction: require('./azurefunction-info'),
   docker: require('./docker-info').getVendorInfo,
   ecs: require('./ecs-info'),
   gcp: require('./gcp-info'),
